tasks 3-4

+++ lms/views.py
@@ -1,3 +1,4 @@
+from rest_framework import status
 from rest_framework.generics import CreateAPIView, DestroyAPIView, ListAPIView, RetrieveAPIView, UpdateAPIView
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
@@ -5,6 +6,7 @@ from rest_framework.views import APIView
 from rest_framework.viewsets import ModelViewSet
 
 from lms.models import Course, Lesson, Subscription
+from lms.paginations import CustomPagination
 from lms.serialilers import CourseDetailSerializer, CourseSerializer, LessonSerializer, SubscriptionSerializer
 from users.permissions import IsModer, IsOwner
 
@@ -12,6 +14,7 @@ from users.permissions import IsModer, IsOwner
 class CourseViewSet(ModelViewSet):
     queryset = Course.objects.all()
     serializer_class = CourseSerializer
+    pagination_class = CustomPagination
 
     def get_serializer_class(self):
         if self.action == "retrieve":
@@ -24,14 +27,12 @@ class CourseViewSet(ModelViewSet):
         course.save()
 
     def get_permissions(self):
-        print(self.action)
         if self.action == "create":
             self.permission_classes = (~IsModer,)
         elif self.action == "destroy":
             self.permission_classes = (~IsModer | IsOwner,)
         elif self.action in ["partial_update", "update", "retrieve"]:
             self.permission_classes = (IsModer | IsOwner,)
-            print(IsModer | IsOwner)
         return super().get_permissions()
 
 
@@ -48,6 +49,7 @@ class LessonCreateAPIView(CreateAPIView):
 class LessonListAPIView(ListAPIView):
     queryset = Lesson.objects.all()
     serializer_class = LessonSerializer
+    pagination_class = CustomPagination
 
 
 class LessonRetrieveAPIView(RetrieveAPIView):
@@ -68,23 +70,20 @@ class LessonDestroyAPIView(DestroyAPIView):
 
 
 class SubscriptionSwitchAPIView(APIView):
-    queryset = Lesson.objects.all()
-    # serializer_class = SubscriptionSerializer
 
     def post(self, *args, **kwargs):
-        user = self.request.user
-        # print (self.request.data)
-        course_id = self.request.data.get("course_id")
-        # print(course_id)
-        course = Course.objects.get(id=course_id)
-        # print(course.name)
-        sub_item = Subscription.objects.filter(user=user, course=course)
-        # print(sub_item)
-        if sub_item.exists():
-            sub_item.delete()
-            message = "Подписка удалена"
+        serializer = SubscriptionSerializer(data=self.request.data)
+        if serializer.is_valid():
+            user = self.request.user
+            course_id = self.request.data.get("course_id")
+            course = Course.objects.get(id=course_id)
+            sub_item = Subscription.objects.filter(user=user, course=course)
+            if sub_item.exists():
+                sub_item.delete()
+                message = "Подписка удалена"
+            else:
+                Subscription.objects.create(user=user, course=course)
+                message = "Подписка добавлена"
+            return Response({"message": message})
         else:
-            Subscription.objects.create(user=user, course=course)
-            message = "Подписка добавлена"
-        print(message)
-        return Response({"message": message})
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
